made reorganization of expression converiers
